add robolectric test in Unit Test, fix exception on local tests for AbstractController, add comp
