+++ app/build.gradle
@@ -45,6 +45,10 @@ dependencies {
     testCompile 'org.slf4j:slf4j-api:1.7.12'
     testCompile 'org.slf4j:slf4j-log4j12:1.7.12'
     testCompile 'com.google.code.gson:gson:2.3.1'
+    testCompile 'org.hamcrest:hamcrest-core:1.1'
+    testCompile 'org.hamcrest:hamcrest-library:1.1'
+    testCompile 'org.hamcrest:hamcrest-integration:1.1'
+    testCompile 'org.robolectric:robolectric:3.0'
     androidTestCompile 'org.mockito:mockito-core:1.10.19'
     androidTestCompile 'com.google.dexmaker:dexmaker-mockito:1.2'
     androidTestCompile 'com.android.support.test:runner:0.3'
